packages/db/pyproject.toml
[project]
name = "meridian-db"
version = "0.1.0"
description = "SQLAlchemy ORM models + Alembic-managed schema for Meridian."
requires-python = ">=3.12"
dependencies = [
    "meridian-contracts",
    "sqlalchemy>=2.0.36",
    "psycopg[binary]>=3.2.3",
    "pgvector>=0.3.6",
    # argon2-cffi: password hashing. argon2id is the OWASP recommendation.
    "argon2-cffi>=23.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meridian_db"]
